+++ plugin/src/main/kotlin/de/infolektuell/gradle/jextract/tasks/DownloadTask.kt
@@ -1,15 +1,12 @@
 package de.infolektuell.gradle.jextract.tasks
 
+import de.infolektuell.gradle.jextract.JextractDataStore
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.file.RegularFileProperty
-import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
-import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.Internal
-import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 import java.io.InputStream
@@ -23,30 +20,17 @@ import java.nio.file.StandardCopyOption
 import java.security.DigestInputStream
 import java.security.MessageDigest
 import java.time.Duration
-import javax.inject.Inject
 
 @CacheableTask
 abstract class DownloadTask : DefaultTask() {
-    abstract class Resource @Inject constructor(objects: ObjectFactory) {
-        @get:Input
-        val url: Property<URI> = objects.property(URI::class.java)
-        @get:Internal
-        val fileName: Provider<String> = url.map { it.path.replaceBeforeLast('/', "").trim('/') }
-        @get:Input
-        abstract val checksum: Property<String>
-        @get:Input
-        abstract val algorithm: Property<String>
-    }
-    @get:Nested
-    abstract val resource: Resource
+    @get:Input
+    abstract val resource: Property<JextractDataStore.Resource>
     @get:OutputFile
     abstract val target: RegularFileProperty
 
     @TaskAction
     protected fun download() {
-        val url = resource.url.get()
-        val checksum = resource.checksum.get()
-        val algorithm = resource.algorithm.get()
+        val (_, url, checksum, algorithm) = resource.get()
         val targetPath = target.get().asFile.toPath()
         Files.createDirectories(targetPath.parent)
         download(url, checksum, algorithm, targetPath)